+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ## DEPENDENCIES
 
 - **chai** has been upgraded from 4.2.0 to **4.3.4**
+- **karma** has been upgraded from 4.4.1 to **6.3.4**
 
 # 6.0.0 / 2020-05-04