+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - **karma** has been upgraded from 4.4.1 to **6.3.4**
 - **karma-mocha** has been upgraded from 1.3.0 to **2.0.1**
 - **mocha** has been upgraded from 7.1.1 to **9.1.1**
+- **video.js** has been upgraded from 7.7.5 to **7.15.4**
 
 # 6.0.0 / 2020-05-04