+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - **chai** has been upgraded from 4.2.0 to **4.3.4**
 - **eslint** has been upgraded from 22.0.0 to **23.0.0**
 - **grunt-eslint** has been removed
+- **grunt-gh-pages** has been removed
 - **grunt-karma** has been removed
 - **grunt-mkdocs** has been removed
 - **karma** has been upgraded from 4.4.1 to **6.3.4**